worker/requirements.txt
redis
psycopg2
compreface-sdk
